+++ packages/components/src/Popover/Popover.test.tsx
@@ -1,6 +1,9 @@
+import '@testing-library/jest-dom/extend-expect'
+import { fireEvent } from '@testing-library/react'
 import 'jest-styled-components'
 import React, { useRef } from 'react'
-import { mountWithTheme } from '@looker/components-test-utils'
+import { renderWithTheme } from '@looker/components-test-utils'
+import { Button } from '../Button'
 import { Popover } from './Popover'
 
 const SimpleContent = <div>simple content</div>
@@ -35,8 +38,17 @@ const PopoverGroup = () => {
 }
 
 describe('Popover', () => {
+  afterEach(() => {
+    const modalRoot = document.getElementById('modal-root')
+    if (modalRoot) {
+      // remove modalRoot after every test so the DOCUMENT_POSITION_FOLLOWING check
+      // won't misleadingly fail since render adds the output to the end of the body
+      document.body.removeChild(modalRoot)
+    }
+  })
+
   test('opens on click', () => {
-    const popover = mountWithTheme(
+    const { getByText, queryByText } = renderWithTheme(
       <Popover content={SimpleContent}>
         {(onClick, ref, className) => (
           <button ref={ref} onClick={onClick} className={className}>
@@ -47,37 +59,22 @@ describe('Popover', () => {
     )
 
     // Verify hidden
-    expect(popover.contains(SimpleContent)).toBeFalsy()
+    expect(queryByText('simple content')).not.toBeInTheDocument()
 
-    const trigger = popover.find('button')
-    trigger.simulate('click')
+    const trigger = getByText('Test')
+    fireEvent.click(trigger)
 
     // Find content
-    expect(popover.contains(SimpleContent)).toBeTruthy()
+    expect(getByText('simple content')).toBeInTheDocument()
 
-    trigger.simulate('click')
-    expect(popover.find(SimpleContent).exists()).toBeFalsy()
-  })
-
-  test('contains the content passed to it', () => {
-    const popover = mountWithTheme(
-      <Popover content={SimpleContent}>
-        {(onClick, ref, className) => (
-          <button ref={ref} onClick={onClick} className={className}>
-            Test
-          </button>
-        )}
-      </Popover>
-    )
-    const trigger = popover.find('button')
-    trigger.simulate('click')
-    expect(popover.contains(SimpleContent)).toBeTruthy()
+    fireEvent.click(trigger)
+    expect(queryByText('simple content')).not.toBeInTheDocument()
   })
 
   test('stopPropagation works - event on container is not called', () => {
     const mockContainerOnClick = jest.fn()
 
-    const popover = mountWithTheme(
+    const { getByText } = renderWithTheme(
       <div onClick={mockContainerOnClick}>
         <Popover content={SimpleContent}>
           {(onClick, ref, className) => (
@@ -88,14 +85,14 @@ describe('Popover', () => {
         </Popover>
       </div>
     )
-    const trigger = popover.find('button')
-    trigger.simulate('click')
-    expect(popover.contains(SimpleContent)).toBeTruthy()
+    const trigger = getByText('Test')
+    fireEvent.click(trigger)
+    expect(getByText('simple content')).toBeInTheDocument()
     expect(mockContainerOnClick).not.toHaveBeenCalled()
   })
 
-  xtest('Popover opens and closes', () => {
-    const popover = mountWithTheme(
+  test('Popover opens and closes', () => {
+    const { getByText, queryByText } = renderWithTheme(
       <Popover content={SimpleContent}>
         {(onClick, ref, className) => (
           <a onClick={onClick} ref={ref} className={className}>
@@ -105,29 +102,23 @@ describe('Popover', () => {
       </Popover>
     )
 
-    const trigger = popover.find('a')
-    expect(trigger.exists()).toBeTruthy()
+    const trigger = getByText('Instant Click')
+    expect(trigger).toBeInTheDocument()
 
     // Verify Popover close, then Open Popover and verify it's open
-    expect(popover.contains(SimpleContent)).toBeFalsy()
-    trigger.simulate('click')
-    expect(popover.contains(SimpleContent)).toBeTruthy()
+    expect(queryByText('simple content')).not.toBeInTheDocument()
+    fireEvent.click(trigger)
+    expect(getByText('simple content')).toBeInTheDocument()
 
     // Collapse Popover by clicking outside of it (original trigger will do)
-    trigger.simulate('click', {})
-
-    /*
-     * Testing this with Jest is frustrating because the component expects to operating with a real DOM
-     * environment. Popover is looking for an event sent via document.addEventListener which isn't
-     * produced within Jest's enviornment. Attempts at mocking haven't been successful.
-     */
-    expect(popover.contains(SimpleContent)).toBeFalsy()
+    fireEvent.click(document)
+    expect(queryByText('simple content')).not.toBeInTheDocument()
   })
 
-  xtest('Open popover cancels event on "dismissal click"', () => {
+  test('Open popover cancels event on "dismissal click"', () => {
     const doThing = jest.fn()
 
-    const popover = mountWithTheme(
+    const { getByText, queryByText } = renderWithTheme(
       <>
         <Popover content={SimpleContent}>
           {(onClick, ref, className) => (
@@ -140,48 +131,77 @@ describe('Popover', () => {
       </>
     )
 
-    const trigger = popover.find('button')
-    trigger.simulate('click') // open Popover
-    const closer = popover.find('a')
-    closer.simulate('click')
+    const trigger = getByText('Instant Click')
+    fireEvent.click(trigger) // open Popover
 
-    /*
-     * Testing this with Jest is frustrating because the component expects to operating with a real DOM
-     * environment. Popover is looking for an event sent via document.addEventListener which isn't
-     * produced within Jest's enviornment. Attempts at mocking haven't been successful.
-     */
-    expect(popover.contains(SimpleContent)).toBeFalsy()
+    const closer = getByText('Do thing...')
+    fireEvent.click(closer)
+    expect(queryByText('simple content')).not.toBeInTheDocument()
     expect(doThing).toBeCalledTimes(0)
   })
 
-  xtest('Popover Group - item outside group does NOT receive first click event', () => {
-    const groupedPopovers = mountWithTheme(<PopoverGroup />)
-    const trigger = groupedPopovers.find('a')
-    trigger.simulate('click') // open Popover
+  test('Popover Group - item outside group does NOT receive first click event', () => {
+    const { getByText, queryByText } = renderWithTheme(<PopoverGroup />)
+    const trigger = getByText('Instant Click')
+    fireEvent.click(trigger) // open Popover
 
-    const dismissed = groupedPopovers.find('button')
-    expect(dismissed.exists()).toBeTruthy()
-    dismissed.simulate('click')
+    const dismissed = getByText('Should require dismissal click')
+    fireEvent.click(dismissed)
     // @FAIL - Doesn't work because Popover is looking for an event sent via document.addEventListener
-    expect(groupedPopovers.contains(SimpleContent)).toBeFalsy()
+    expect(queryByText('simple content')).not.toBeInTheDocument()
     expect(requiresDismissal).toBeCalledTimes(0)
   })
 
-  xtest('Popover Group  - item within group immediately receives onClick', () => {
-    const groupedPopovers = mountWithTheme(<PopoverGroup />)
-    const trigger = groupedPopovers.find('a')
-    trigger.simulate('click') // open Popover
+  test('Popover Group  - item within group immediately receives onClick', () => {
+    const { getByText, queryByText } = renderWithTheme(<PopoverGroup />)
+    const trigger = getByText('Instant Click')
+    fireEvent.click(trigger) // open Popover
 
-    const instant = groupedPopovers.find('#instant')
-    expect(instant.exists()).toBeTruthy()
-    instant.simulate('click')
+    const instant = getByText('Should activate instantly')
+    fireEvent.click(instant)
 
     /*
      * Testing this with Jest is frustrating because the component expects to operating with a real DOM
      * environment. Popover is looking for an event sent via document.addEventListener which isn't
      * produced within Jest's enviornment. Attempts at mocking haven't been successful.
      */
-    expect(groupedPopovers.contains(SimpleContent)).toBeFalsy()
+    expect(queryByText('simple content')).not.toBeInTheDocument()
     expect(instantClick).toBeCalledTimes(1)
+  })
+
+  test('Popover trigger is shown/hidden on hover of hoverDisclosureRef', () => {
+    const Component = () => {
+      const hoverRef = React.useRef<HTMLDivElement>(null)
+      return (
+        <div ref={hoverRef}>
+          <Popover content={SimpleContent} hoverDisclosureRef={hoverRef}>
+            {(onClick, ref, className) => (
+              <Button onClick={onClick} ref={ref} className={className}>
+                Instant Click
+              </Button>
+            )}
+          </Popover>
+          Some text in the div
+        </div>
+      )
+    }
+    const { getByText, queryByText } = renderWithTheme(<Component />)
+
+    const trigger = queryByText('Instant Click')
+    const div = getByText('Some text in the div')
+
+    expect(trigger).not.toBeInTheDocument()
+
+    fireEvent(
+      div,
+      new MouseEvent('mouseenter', {
+        bubbles: true,
+        cancelable: true,
+      })
+    )
+    const triggerNew = getByText('Instant Click')
+    expect(queryByText('simple content')).not.toBeInTheDocument()
+    fireEvent.click(triggerNew) // open Popover
+    expect(queryByText('simple content')).toBeInTheDocument()
   })
 })